Migrate any float to double.

+++ jplag.cli/src/test/java/de/jplag/cli/CommandLineInterfaceTest.java
@@ -19,7 +19,7 @@ import de.jplag.options.JPlagOptions;
  */
 public abstract class CommandLineInterfaceTest {
     protected static final String CURRENT_DIRECTORY = ".";
-    protected static final float DELTA = 0.0001f;
+    protected static final double DELTA = 0.0001f;
 
     protected CLI cli;
     protected Namespace namespace;
